clean up man utils
